beginning testing of new energyeval
