+++ src/EnergyEvaluation.jl
@@ -1,33 +1,27 @@
 """
     module EnergyEvaluation
+Rewrite of the original energyevaluation module to standardise and clean the functions up. 
+"""
 
-this module provides data, structs and methods for dimer energy and total energy evaluation
-"""    
-module EnergyEvaluation
+module EnergyEvaluation 
 
-using StaticArrays 
-using DFTK 
-using LinearAlgebra
-using SplitApplyCombine
+using StaticArrays,LinearAlgebra#,DFTK
+
 using ..MachineLearningPotential
 using ..Configurations
 
-export AbstractPotential, AbstractDimerPotential, AbstractDimerPotentialB, AbstractMachineLearningPotential,EmbeddedAtomPotential
+export AbstractEnsemble,NVT,NPT
+export AbstractPotential,AbstractDimerPotential,ELJPotential,ELJPotentialEven
+export AbstractDimerPotentialB,ELJPotentialB,EmbeddedAtomPotential,RuNNerPotential
+export PotentialVariables,DimerPotentialVariables,ELJPotentialBVariables
+export EmbeddedAtomVariables,NNPVariables
 
+export energy_update!,set_variables,initialise_energy
+#-------------------------------------------------------------#
+#----------------------Universal Structs----------------------#
+#-------------------------------------------------------------#
 
-export RuNNerPotential
-
-export DFTPotential
-
-export ELJPotential, ELJPotentialEven, ELJPotentialB
-export dimer_energy, dimer_energy_atom, dimer_energy_config 
-export getenergy_DFT, get_energy_dimer#,get_energy_
-export energy_update
-export get_energy
-export AbstractEnsemble, NVT, NPT
-export EnHist
-
-
+#--------------------------Ensembles--------------------------#
 """
     AbstractEnsemble
 abstract type for ensemble:
@@ -57,6 +51,7 @@ struct NPT <: AbstractEnsemble
     pressure::Real
 end
 
+#-------------------------------------------------------------#
 """   
     AbstractPotential
 Abstract type for possible potentials
@@ -68,6 +63,16 @@ Needs method for dimer_energy [`dimer_energy`](@ref)
 abstract type AbstractPotential end
 
 """
+    PotentialVariables
+An abstract type defining a class of mutable struct containing all the relevant vectors and arrays each potential will need throughout the course of a simulation to prevent over-definitions inside the MCState struct.
+"""
+abstract type PotentialVariables end
+#-----------------------------------------------------------------------#
+#-----------------------Explicit Dimer Potentials-----------------------#
+#-----------------------------------------------------------------------#
+
+#--------------------------Dimer Structs--------------------------#
+"""
     AbstractDimerPotential <: AbstractPotential
  implemented dimer potentials:   
     - ELJPotential [`ELJPotential`](@ref)
@@ -78,35 +83,26 @@ Needs methods for
     - dimer_energy_config [`dimer_energy_config`](@ref)
 """   
 abstract type AbstractDimerPotential <: AbstractPotential end
-
-#The potential struct for magnetic field (tan(theta) as variable)
-abstract type AbstractDimerPotentialB <: AbstractPotential end
-
-
-abstract type AbstractMachineLearningPotential <: AbstractPotential end
-"""
-    PotentialVariables
-An abstract type defining a class of mutable struct containing all the relevant vectors and arrays each potential will need throughout the course of a simulation to prevent over-definitions inside the MCState struct.
-"""
-abstract type PotentialVariables end
-
-#-----------------------------------------------------------------------#
-#-----------------------Explicit Dimer Potentials-----------------------#
-#-----------------------------------------------------------------------#
 """
     DimerPotenitalVariables
 The struct contains only the new_dist2_vec as this doesn't explicitly require any particular special features.
 """
 mutable struct DimerPotentialVariables <: PotentialVariables
-    new_dist2_vec::Vector
-    new_en::Float64
+    #new_dist2_vec::Vector
+    en_atom_vec::Vector
+    #new_en::Float64
 end
+
 """
     dimer_energy_atom(i, pos, d2vec, pot<:AbstractPotential)
-Sums the dimer energies for atom `i` with all other atoms
+    dimer_energy_atom(i, d2vec, r_cut, pot:abstract type AbstractDimerPotential <: AbstractPotential end:AbstractDimerPotential)
+
+Sums the dimer energies for atom `i` with all other atomsdimer_energy_update(pos,index,config,dist2_mat,new_dist2_vec,en_tot,pot::AbstractDimerPotential)
 Needs vector of squared distances `d2vec` between atom `i` and all other atoms in configuration
 see  `get_distance2_mat` [`get_distance2_mat`](@ref) 
 and potential information `pot` [`Abstract_Potential`](@ref) 
+
+second method includes r_cut for use with the NPT ensemble
 """
 function dimer_energy_atom(i, d2vec, pot::AbstractDimerPotential)
     sum1 = 0.
@@ -119,28 +115,9 @@ function dimer_energy_atom(i, d2vec, pot::AbstractDimerPotential)
     return sum1
 end
 
-
-"""
-    dimer_energy_atom(i, d2vec, r_cut, pot<:AbstractPotential)
-    dimer energy of an atom with other atoms, with a cutoff distance r_cut.
-"""
-function dimer_energy_atom(i, d2vec, r_cut, pot::AbstractDimerPotential)
-    sum1 = 0.
-    for j in 1:i-1
-        if d2vec[j] <= r_cut
-            sum1 += dimer_energy(pot, d2vec[j])
-        end
-    end
-    for j in i+1:size(d2vec,1)
-        if d2vec[j] <= r_cut
-            sum1 += dimer_energy(pot, d2vec[j])
-        end
-    end 
-    return sum1
-end
-
 """
     dimer_energy_config(distmat, NAtoms, pot::AbstractPotential)
+    dimer_energy_config(distmat, NAtoms, r_cut, pot::AbstractPotential)
 Stores the total of dimer energies of one atom with all other atoms in vector and
 calculates total energy of configuration
 Needs squared distances matrix, see `get_distance2_mat` [`get_distance2_mat`](@ref) 
@@ -149,11 +126,7 @@ and potential information `pot` [`Abstract_Potential`](@ref)
 function dimer_energy_config(distmat, NAtoms, pot::AbstractDimerPotential)
     dimer_energy_vec = zeros(NAtoms)
     energy_tot = 0.
-    #This uses dimer_energy_atom for all atoms, which means one dimer energy will be calculated twice.
-    #for i in 1:NAtoms #eachindex(),enumerate()..?
-        #dimer_energy_vec[i] = dimer_energy_atom(i, distmat[:, i], pot) #@view distmat[i, :]
-        #energy_tot += dimer_energy_vec[i]
-    #end 
+
     for i in 1:NAtoms
         for j=i+1:NAtoms
             e_ij=dimer_energy(pot,distmat[i,j])
@@ -165,146 +138,18 @@ function dimer_energy_config(distmat, NAtoms, pot::AbstractDimerPotential)
     #energy_tot=sum(dimer_energy_vec)
     return dimer_energy_vec, energy_tot
 end 
-
-
 """
-    dimer_energy_config(distmat, NAtoms, r_cut, pot::AbstractPotential)
-Stores the total of dimer energies of one atom with all other atoms in vector and
-calculates total energy of configuration, with a cutoff distance r_cut
-Needs squared distances matrix, see `get_distance2_mat` [`get_distance2_mat`](@ref) 
-and potential information `pot` [`Abstract_Potential`](@ref) 
-"""
-function dimer_energy_config(distmat, NAtoms, r_cut, pot::AbstractDimerPotential)
-    dimer_energy_vec = zeros(NAtoms)
-    energy_tot = 0.
-    #for i in 1:NAtoms #eachindex(),enumerate()..?
-        #dimer_energy_vec[i] = dimer_energy_atom(i, distmat[:, i], r_cut, pot) #@view distmat[i, :]
-        #energy_tot += dimer_energy_vec[i]
-    #end 
+    dimer_energy_update(pos,index,config,dist2_mat,new_dist2_vec,en_tot,pot::AbstractDimerPotential)
+dimer_energy_update is the potential-level-call where for a single mc_state we take the new position `pos`, for atom at `index` , inside the current `config` , where the interatomic distances `dist2_mat` and the new vector based on the new position `new_dist2_vec`; these use the `potential` to calculate a delta_energy and modify the current `en_tot`. These quantities are modified in place and returned 
+""" 
+function dimer_energy_update!(index,dist2_mat,new_dist2_vec,en_tot,pot::AbstractDimerPotential)
+    delta_en = dimer_energy_atom(index,new_dist2_vec,pot) - dimer_energy_atom(index,dist2_mat[index,:], pot)
 
-    for i in 1:NAtoms
-        for j=i+1:NAtoms
-            if distmat[i,j] <= r_cut
-                e_ij=dimer_energy(pot,distmat[i,j])
-                dimer_energy_vec[i] += e_ij
-                dimer_energy_vec[j] += e_ij
-                energy_tot += e_ij
-            end
-        end
-    end 
-
-    return dimer_energy_vec, energy_tot + lrc(NAtoms,r_cut,pot)   #no 0.5*energy_tot
-end    
-
-"""
-    energy_update(i_atom, dist2_new, pot::AbstractDimerPotential)
-Gives new energy when an atom is moved. Without cutoff ditance.
-"""
-function energy_update(i_atom, dist2_new, pot::AbstractDimerPotential)
-    return dimer_energy_atom(i_atom, dist2_new, pot)
+    return  delta_en + en_tot
 end
-
-"""
-    energy_update(i_atom, dist2_new, r_cut, pot::AbstractDimerPotential)
-Gives new energy when an atom is moved. With cutoff distance r_cut
-"""
-function energy_update(i_atom, dist2_new, r_cut, pot::AbstractDimerPotential)
-    return dimer_energy_atom(i_atom, dist2_new, r_cut, pot)
-end
-
-"""
-    energy_update(pos, i_atom, config, dist2_mat, pot::AbstractDimerPotential)
-Gives energy difference and distances when an atom is moved. Without cutoff distance.
-"""
-function energy_update(pos, i_atom, config, dist2_mat, pot::AbstractDimerPotential)
-    dist2_new = [distance2(pos,b,config.bc) for b in config.pos]
-    dist2_new[i_atom] = 0.
-    d_en = dimer_energy_atom(i_atom, dist2_new, pot) - dimer_energy_atom(i_atom, dist2_mat[:,i_atom], pot)
-    return d_en, dist2_new
-end
-
-"""
-    energy_update(pos, i_atom, config, dist2_mat, r_cut, pot::AbstractDimerPotential)
-Gives energy difference and distances when an atom is moved. With cutoff distance r_cut.
-"""
-function energy_update(pos, i_atom, config, dist2_mat, r_cut, pot::AbstractDimerPotential)
-    dist2_new = [distance2(pos,b,config.bc) for b in config.pos]
-    dist2_new[i_atom] = 0.
-    d_en = dimer_energy_atom(i_atom, dist2_new, r_cut, pot) - dimer_energy_atom(i_atom, dist2_mat[:,i_atom], r_cut, pot)
-    return d_en, dist2_new
-end
-"""
-    get_energy_dimer(pos,i_atom,mc_state,pot)
-A get_energy function similar to the energy_update function. This simply returns the current energy rather than delta_en
-"""
-function get_energy_dimer(pos,i_atom,mc_state,pot::AbstractDimerPotential)
-
-    delta_energy,dist2_new = energy_update(pos,i_atom,mc_state.config,mc_state.dist2_mat,pot)
-    energy = mc_state.en_tot + delta_energy
-    return energy,dist2_new
-end
-"""
-    get_energy_dimer(pos,i_atom,r_cut,mc_state,pot)
-A get_energy function similar to the energy_update function. This simply returns the current energy rather than delta_en
-with a cutoff distance r_cut
-"""
-function get_energy_dimer(pos,i_atom,r_cut,mc_state,pot::AbstractDimerPotential)
-    # dist2_new = [distance2(pos,b) for b in mc_state.config.pos]
-    # dist2_new[i_atom] = 0.
-    # delta_energy= dimer_energy_atom(i_atom, dist2_new, pot) - dimer_energy_atom(i_atom, mc_state.dist2_mat[:,i_atom], pot)
-
-    delta_energy,dist2_new = energy_update(pos,i_atom,mc_state.config,mc_state.dist2_mat,r_cut,pot)
-    energy = mc_state.en_tot + delta_energy
-    return energy,dist2_new
-end
-
-"""
-    get_energy(trial_positions,indices,mc_states,pot::AbstractDimerPotential)
-Top scope get energy function for dimer potentials returning the energy vector and new distance squared vector as this must be calculated in order to calculate the potential.
-"""
-function get_energy(trial_positions,indices,mc_states,pot::AbstractDimerPotential,ensemble::NVT)
-    energyvector, dist2_new = invert(get_energy_dimer.(trial_positions,indices,mc_states,Ref(pot)))
-    # energyvector = mc_state.en_tot .+ delta_energyvector
-    return energyvector,dist2_new
-end
-#this will be the format for this part of the get_energy function.
-
-"""
-    get_energy(trial_positions,indices,r_cut,mc_states,pot::AbstractDimerPotential,ensemble::NPT)
-Top scope get energy function for dimer potentials returning the energy vector and new distance squared vector as this must be calculated in order to calculate the potential.
-with the cutoff distance r_cut
-"""
-function get_energy(trial_positions,indices,mc_states,pot::AbstractDimerPotential,ensemble::NPT)
-    
-    n=length(mc_states)
-    r_cut_all=Array{Float64}(undef,n)
-    for i=1:n
-        r_cut_all[i]=mc_states[i].config.bc.box_length^2/4
-    end
-    energyvector, dist2_new = invert(get_energy_dimer.(trial_positions,indices,r_cut_all,mc_states,Ref(pot)))
-    # energyvector = mc_state.en_tot .+ delta_energyvector
-    return energyvector,dist2_new
-end
-
-"""
-    get_energy function when the whole configuration is scaled
-    find the new distance matrix first, and use dimer_energy_config to find the new total energy and energy matrix
-"""
-function get_energy(trial_configs_all,mc_states,pot::AbstractDimerPotential)
-    dist2_mat_new = get_distance2_mat.(trial_configs_all)
-    #en_atom_vec, en_tot_new = invert(dimer_energy_config.(dist2_mat_new, length(trial_configs_all[1].pos), Ref(pot)))
-    n=length(trial_configs_all)
-    r_cut_all=Array{Float64}(undef,n)
-    for i=1:n
-        r_cut_all[i]=trial_configs_all[i].bc.box_length^2/4
-    end
-    en_atom_vec, en_tot_new = invert(dimer_energy_config.(dist2_mat_new, length(trial_configs_all[1].pos), r_cut_all, Ref(pot)))
-    #println(invert(dimer_energy_config.(dist2_mat_new, length(trial_configs_all[1].pos), Ref(pot))))
-    #println()
-    return dist2_mat_new,en_atom_vec,en_tot_new
-end
-
-# energyvector, dist2new = invert(get_energy_dimer.(trial_positions,indices,mc_states,Ref(pot)))
+#----------------------------------------------------------#
+#-----------------Specific Dimer Potentials----------------#
+#----------------------------------------------------------#
 """
     ELJPotential{N,T} 
 Implements type for extended Lennard Jones potential; subtype of [`AbstractDimerPotential`](@ref)<:[`AbstractPotential`](@ref);
@@ -382,9 +227,114 @@ function dimer_energy(pot::ELJPotentialEven{N}, r2) where N
     end
     return sum1
 end 
-#------------------------------------------------------------------------#
-#-------------------------EmbeddedAtomPotential--------------------------#
-#------------------------------------------------------------------------#
+#----------------------------------------------------------#
+
+#-----------------Magnetic Dimer Potential-----------------#
+abstract type AbstractDimerPotentialB <: AbstractPotential end
+"""
+   potential in B
+"""
+struct ELJPotentialB{N,T} <: AbstractDimerPotentialB
+    coeff_a::SVector{N,T}
+    coeff_b::SVector{N,T}
+    coeff_c::SVector{N,T}
+end
+function ELJPotentialB{N}(a,b,c) where N
+    @boundscheck length(c) == N || error("number of ELJ coefficients does not match given length")
+    coeff_a = SVector{N}(a)
+    coeff_b = SVector{N}(b)
+    coeff_c = SVector{N}(c)
+    T = eltype(c)
+    return ELJPotentialB{N,T}(coeff_a,coeff_b,coeff_c)
+end
+
+function ELJPotentialB(a,b,c) 
+    N = length(c)
+    coeff_a = SVector{N}(a)
+    coeff_b = SVector{N}(b)
+    coeff_c = SVector{N}(c)
+    T = eltype(c)
+    return ELJPotentialB{N,T}(coeff_a,coeff_b,coeff_c)
+end
+
+mutable struct ELJPotentialBVariables <: PotentialVariables
+    en_atom_vec::Array
+    tan_mat::Array
+    new_tan_vec::Vector
+end
+"""
+    dimer_energy(pot::ELJPotentialB{N}, r2, tan) where N
+Dimer energy when the distance square between two atom is r2 and the angle between the line connecting them and z-direction is tan.
+When r2 < 5.30, returns 1.
+"""
+function dimer_energy(pot::ELJPotentialB{N}, r2, tan) where N
+   
+    if r2>=5.30
+        r6inv = 1/(r2*r2*r2)
+        t2=2/(tan^2+1)-1     #cos(2*theta)
+        t4=2*t2^2-1
+        sum1 = pot.coeff_c[1] * r6inv * (1 + pot.coeff_a[1]*t2 + pot.coeff_b[1]*t4)
+        r6inv/=r2
+        for i = 2:N
+            sum1 += pot.coeff_c[i] * r6inv * (1 + pot.coeff_a[i]*t2 + pot.coeff_b[i]*t4)
+            r6inv /= r2^0.5 
+        end
+    else
+        sum1=0.1
+    end
+    return sum1
+end 
+
+"""
+    dimer_energy_atom(i, pos, d2vec, pot<:AbstractDimerPotentialB)
+Sums the dimer energies for atom `i` with all other atoms
+Needs vector of squared distances `d2vec` between atom `i` and all other atoms in configuration
+see  `get_distance2_mat` [`get_distance2_mat`](@ref) 
+and potential information `pot` [`Abstract_Potential`](@ref) 
+"""
+function dimer_energy_atom(i, d2vec, tanvec,pot::AbstractDimerPotentialB)
+    sum1 = 0.
+    for j in 1:i-1
+        sum1 += dimer_energy(pot, d2vec[j], tanvec[j])
+    end
+    for j in i+1:size(d2vec,1)
+        sum1 += dimer_energy(pot, d2vec[j], tanvec[j])
+    end 
+    return sum1
+end
+
+"""
+    dimer_energy_config(distmat, NAtoms, pot::AbstractDimerPotentialB)
+Stores the total of dimer energies of one atom with all other atoms in vector and
+calculates total energy of configuration
+Needs squared distances matrix, see `get_distance2_mat` [`get_distance2_mat`](@ref) 
+and potential information `pot` [`Abstract_Potential`](@ref) 
+"""
+function dimer_energy_config(distmat, tanmat, NAtoms, pot::AbstractDimerPotentialB)
+    dimer_energy_vec = zeros(NAtoms)
+    energy_tot = 0.
+
+    for i in 1:NAtoms
+        for j=i+1:NAtoms
+            e_ij=dimer_energy(pot,distmat[i,j],tanmat[i,j])
+            dimer_energy_vec[i] += e_ij
+            dimer_energy_vec[j] += e_ij
+            energy_tot += e_ij
+        end
+    end 
+    #energy_tot=sum(dimer_energy_vec)
+    return dimer_energy_vec, energy_tot
+end 
+function dimer_energy_update!(index,dist2_mat,tanmat,new_dist2_vec,new_tan_vec,en_tot,pot::AbstractDimerPotentialB)
+    # new_dist2_vec = [distance2(pos,b,config.bc) for b in config.pos]
+    # new_dist2_vec[index] = 0.
+    delta_en = dimer_energy_atom(index,new_dist2_vec,new_tan_vec,pot) - dimer_energy_atom(index,dist2_mat[index,:],tanmat[index,:], pot)
+
+    return  delta_en + en_tot
+end
+#----------------------------------------------------------#
+#-------------------Embedded Atom Model--------------------#
+#----------------------------------------------------------#
 """
     EmbeddedAtomPotential
 Struct containing the important quantities for calculating EAM (specifically Sutton-Chen type) potentials.
@@ -410,6 +360,12 @@ function EmbeddedAtomPotential(n,m,ϵ,C,a)
     epsCam = ϵ * C * a^(m/2)
     return EmbeddedAtomPotential(n,m,epsan,epsCam)
 end
+
+mutable struct EmbeddedAtomVariables <: PotentialVariables
+    component_vector::Matrix
+    new_component_vector::Matrix
+end
+#-------------------Component Calculation------------------#
 """
     invrexp(r2,n,m)
 Function to calculate the `ϕ,ρ` components given a square distance `r2` and the exponents `n,m`
@@ -424,32 +380,19 @@ function invrexp(r2,n,m)
 end
 """
     calc_components(eatomvec,distancevec,n,m)
-Primary calculation of ϕ,ρ for atom i, given each other atom's distance to i in `distancevec`. `eatomvec` stores the ϕ and ρ components.
-"""
-function calc_components(eatomvec,distancevec,n,m)
-    for dist in distancevec
-        eatomvec .+= invrexp(dist,n,m)
-    end
-    return eatomvec
-end
-"""
-    calc_components(dist2_mat,n,m) 
-    calc_components(componentvec,atomindex,old_r2_vec,new_r2_vec,n,m)
-Given a matrix `dist2_mat` of each square-radial distance and the exponents `n,m` calculates the Nx2 vector of ϕ and ρ components for each atom and returning them as a `component_vector`.
+    calc_components(new_component_vec,atomindex,old_r2_vec,new_r2_vec,n,m)
 
-Second method also includes an existing `componentvec` `atomindex` and old and new interatomic distances from an atom at `atomindex` stored in vectors `new_r2_vec,old_r2_vec`. This calculates the `new_component_vec` based on the updated distances and returns this.
+Primary calculation of ϕ,ρ for atom i, given each other atom's distance to i in `distancevec`. `eatomvec` stores the ϕ and ρ components.
+
+Second method also includes an existing `new_component_vec` `atomindex` and old and new interatomic distances from an atom at `atomindex` stored in vectors `new_r2_vec,old_r2_vec`. This calculates the `new_component_vec` based on the updated distances and returns this.
 """
-function calc_components(dist2_mat,n,m)
-    n_atoms = size(dist2_mat)[1]
-    component_vector = zeros(n_atoms,2)
-    for row_index in 1:n_atoms
-        component_vector[row_index,:] = calc_components(component_vector[row_index,:],dist2_mat[row_index,:],n,m)
+function calc_components(componentvec,distancevec,n,m)
+    for dist in distancevec
+        componentvec .+= invrexp(dist,n,m)
     end
-    return component_vector
+    return componentvec
 end
 function calc_components(new_component_vec,atomindex,old_r2_vec,new_r2_vec,n,m)
-
-    #new_component_vec = copy(componentvec)
 
     for j_index in eachindex(new_r2_vec)
 
@@ -458,6 +401,7 @@ function calc_components(new_component_vec,atomindex,old_r2_vec,new_r2_vec,n,m)
         new_component_vec[j_index,:] .+= j_term 
         new_component_vec[atomindex,:] .+= j_term 
     end
+
     return new_component_vec
 end
 """
@@ -467,421 +411,12 @@ Takes a `component_vector` containing ϕ,ρ for each atom. Using the multiplicat
 function calc_energies_from_components(component_vector,ean,ecam)
     return sum(ean.*component_vector[:,1] - ecam*sqrt.(component_vector[:,2]))
 end
-"""
-    calc_new_energy(componentvec,atomindex,old_r2,new_r2,pot::EmbeddedAtomPotential)
-Function to calculate updated energy after a move of an atom at `atomindex` resulting in a change in `componentvec` to `new_component_vector`. The `dist2_mat` and `newpos` are used with the potential `pot` to calculate the updated components and `newenergy` 
-"""
-function calc_new_energy(mc_state,atomindex,newpos,pot::EmbeddedAtomPotential)
+#-----------------------------------------------------------#
+#----------------Machine Learning Potentials----------------#
+#-----------------------------------------------------------#
 
-    dist2_new = [distance2(newpos,b) for b in mc_state.config.pos]
-    dist2_new[atomindex] = 0.
-    new_component_vector = deepcopy(mc_state.en_atom_vec)
-    new_component_vector = calc_components(new_component_vector,atomindex,mc_state.dist2_mat[atomindex,:],dist2_new,pot.n,pot.m)
+abstract type AbstractMachineLearningPotential <: AbstractPotential end
 
-    newenergy = calc_energies_from_components(new_component_vector,pot.ean,pot.eCam)
-
-    return  newenergy,dist2_new,new_component_vector 
-end
-
-"""
-    get_energy(dist2_mat,pot::EmbeddedAtomPotential)
-Initialises the energy of an EAM potential. Using the interatomic distances stored in `dist2_mat` and the parameters in `pot` we return the atomic two body components in `componentvec` as well as the total energy. 
-"""
-function get_energy(dist2_mat,pot::EmbeddedAtomPotential)
-    componentvec = calc_components(dist2_mat,pot.n,pot.m)
-    return componentvec, calc_energies_from_components(componentvec,pot.ean,pot.eCam) 
-end
-
-function get_energy(trial_positions,indices,mc_states,pot::EmbeddedAtomPotential)
-
-    energyvector,dist2_new,new_components = invert(calc_new_energy.(mc_states,indices,trial_positions,Ref(pot)))
-
-
-
-    return energyvector,dist2_new,new_components
-end
-"""
-    dimer_energy_config(distmat,NAtoms,pot::EmbeddedAtomPotential)
-Currying function to match the initialisation of MCStates with the embedded atom model. This now matches the form of the general dimer potential for use.
-"""
-function dimer_energy_config(distmat,NAtoms,pot::EmbeddedAtomPotential)
-    
-    return get_energy(distmat,pot)
-end
-#------------------------------------------------------------------------#
-#--------------------------Non-Dimer Potentials--------------------------#
-#------------------------------------------------------------------------#
-
-"""
-    lrc(NAtoms,r_cut,pot::ELJPotentialEven{N})
-The long range correction for the extended Lennard-Jones potential (even). r_cut is the cutoff distance.
-lrc is an integral of all interactions outside the cutoff distance, using the uniform density approximation.
-"""
-function lrc(NAtoms,r_cut,pot::ELJPotentialEven{N}) where N
-    
-    r_cut_sqrt=r_cut^0.5
-    rc3 = r_cut*r_cut_sqrt
-    e_lrc = 0.
-    for i = 1:N
-        e_lrc += pot.coeff[i] / rc3 / (2i+1)
-        rc3 *= r_cut
-    end
-    e_lrc *= pi*NAtoms^2/4/r_cut_sqrt^3
-    return e_lrc
-end
-
-
-
-"""
-   potential in B
-"""
-struct ELJPotentialB{N,T} <: AbstractDimerPotentialB
-    coeff_a::SVector{N,T}
-    coeff_b::SVector{N,T}
-    coeff_c::SVector{N,T}
-end
-
-function ELJPotentialB{N}(a,b,c) where N
-    @boundscheck length(c) == N || error("number of ELJ coefficients does not match given length")
-    coeff_a = SVector{N}(a)
-    coeff_b = SVector{N}(b)
-    coeff_c = SVector{N}(c)
-    T = eltype(c)
-    return ELJPotentialB{N,T}(coeff_a,coeff_b,coeff_c)
-end
-
-function ELJPotentialB(a,b,c) 
-    N = length(c)
-    coeff_a = SVector{N}(a)
-    coeff_b = SVector{N}(b)
-    coeff_c = SVector{N}(c)
-    T = eltype(c)
-    return ELJPotentialB{N,T}(coeff_a,coeff_b,coeff_c)
-end
-
-"""
-    dimer_energy(pot::ELJPotentialB{N}, r2, tan) where N
-Dimer energy when the distance square between two atom is r2 and the angle between the line connecting them and z-direction is tan.
-When r2 < 5.30, returns 1.
-"""
-function dimer_energy(pot::ELJPotentialB{N}, r2, tan) where N
-    #d^(-6)*c6*(1+a6*t2+b6*t4)+d^(-8)*(c8*(1+a8*t2+b8*t4)+d_inv*(c9*(1+a9*t2+b9*t4)+d_inv*(c10*(1+a10*t2+b10*t4)+d_inv*(c11*(1+a11*t2+b11*t4)+d_inv*c12*(1+a12*t2+b12*t4)))))
-    if r2>=5.30
-        r6inv = 1/(r2*r2*r2)
-        t2=2/(tan^2+1)-1     #cos(2*theta)
-        t4=2*t2^2-1
-        sum1 = pot.coeff_c[1] * r6inv * (1 + pot.coeff_a[1]*t2 + pot.coeff_b[1]*t4)
-        r6inv/=r2
-        for i = 2:N
-            sum1 += pot.coeff_c[i] * r6inv * (1 + pot.coeff_a[i]*t2 + pot.coeff_b[i]*t4)
-            r6inv /= r2^0.5 
-        end
-    else
-        sum1=0.1
-    end
-    return sum1
-end 
-
-
-"""
-    lrc(NAtoms,r_cut,pot::ELJPotentialEven{N})
-The long range correction for the energy in magnetic field. r_cut is the cutoff distance.
-lrc is an integral of all interactions outside the cutoff distance, using the uniform density approximation.
-At long range, dimer energy changes very little with tan(theta), the coefficients are fitted to energies from theta = 0 to 90 degrees.
-"""
-function lrc(NAtoms,r_cut,pot::ELJPotentialB{N}) where N
-    coeff=[-0.1279111890228638, -1.328138539967966, 12.260941135261255,41.12212408251662]
-    r_cut_sqrt=r_cut^0.5
-    rc3 = r_cut*r_cut_sqrt
-    e_lrc = 0.
-    for i = 1:4
-        e_lrc += coeff[i] / rc3 / (2i+1)
-        rc3 *= r_cut
-    end
-    e_lrc *= pi*NAtoms^2/4/r_cut_sqrt^3
-    return e_lrc
-end
-
-"""
-    all associated energy calculation function with ELJ potential (with B)
-"""
-
-"""
-    dimer_energy_atom(i, pos, d2vec, pot<:AbstractDimerPotentialB)
-Sums the dimer energies for atom `i` with all other atoms
-Needs vector of squared distances `d2vec` between atom `i` and all other atoms in configuration
-see  `get_distance2_mat` [`get_distance2_mat`](@ref) 
-and potential information `pot` [`Abstract_Potential`](@ref) 
-"""
-function dimer_energy_atom(i, d2vec, tanvec,pot::AbstractDimerPotentialB)
-    sum1 = 0.
-    for j in 1:i-1
-        sum1 += dimer_energy(pot, d2vec[j], tanvec[j])
-    end
-    for j in i+1:size(d2vec,1)
-        sum1 += dimer_energy(pot, d2vec[j], tanvec[j])
-    end 
-    return sum1
-end
-
-
-"""
-    dimer_energy_atom(i, d2vec, r_cut, pot<:AbstractDimerPotentialB)
-    dimer energy of an atom with other atoms, with a cutoff distance r_cut.
-"""
-function dimer_energy_atom(i, d2vec, tanvec, r_cut, pot::AbstractDimerPotentialB)
-    sum1 = 0.
-    for j in 1:i-1
-        if d2vec[j] <= r_cut
-            sum1 += dimer_energy(pot, d2vec[j], tanvec[j])
-        end
-    end
-    for j in i+1:size(d2vec,1)
-        if d2vec[j] <= r_cut
-            sum1 += dimer_energy(pot, d2vec[j], tanvec[j])
-        end
-    end 
-    return sum1
-end
-
-"""
-    dimer_energy_config(distmat, NAtoms, pot::AbstractDimerPotentialB)
-Stores the total of dimer energies of one atom with all other atoms in vector and
-calculates total energy of configuration
-Needs squared distances matrix, see `get_distance2_mat` [`get_distance2_mat`](@ref) 
-and potential information `pot` [`Abstract_Potential`](@ref) 
-"""
-function dimer_energy_config(distmat, tanmat, NAtoms, pot::AbstractDimerPotentialB)
-    dimer_energy_vec = zeros(NAtoms)
-    energy_tot = 0.
-    #This uses dimer_energy_atom for all atoms, which means one dimer energy will be calculated twice.
-    #for i in 1:NAtoms #eachindex(),enumerate()..?
-        #dimer_energy_vec[i] = dimer_energy_atom(i, distmat[:, i], pot) #@view distmat[i, :]
-        #energy_tot += dimer_energy_vec[i]
-    #end 
-    for i in 1:NAtoms
-        for j=i+1:NAtoms
-            e_ij=dimer_energy(pot,distmat[i,j],tanmat[i,j])
-            dimer_energy_vec[i] += e_ij
-            dimer_energy_vec[j] += e_ij
-            energy_tot += e_ij
-        end
-    end 
-    #energy_tot=sum(dimer_energy_vec)
-    return dimer_energy_vec, energy_tot
-end 
-
-
-"""
-    dimer_energy_config(distmat, NAtoms, r_cut, pot::AbstractDimerPotentialB)
-Stores the total of dimer energies of one atom with all other atoms in vector and
-calculates total energy of configuration, with a cutoff distance r_cut
-Needs squared distances matrix, see `get_distance2_mat` [`get_distance2_mat`](@ref) 
-and potential information `pot` [`Abstract_Potential`](@ref) 
-"""
-function dimer_energy_config(distmat, tanmat, NAtoms, r_cut, pot::AbstractDimerPotentialB)
-    dimer_energy_vec = zeros(NAtoms)
-    energy_tot = 0.
-    #for i in 1:NAtoms #eachindex(),enumerate()..?
-        #dimer_energy_vec[i] = dimer_energy_atom(i, distmat[:, i], r_cut, pot) #@view distmat[i, :]
-        #energy_tot += dimer_energy_vec[i]
-    #end 
-
-    for i in 1:NAtoms
-        for j=i+1:NAtoms
-            if distmat[i,j] <= r_cut
-                e_ij=dimer_energy(pot,distmat[i,j],tanmat[i,j])
-                dimer_energy_vec[i] += e_ij
-                dimer_energy_vec[j] += e_ij
-                energy_tot += e_ij
-            end
-        end
-    end 
-
-    return dimer_energy_vec, energy_tot + lrc(NAtoms,r_cut,pot)   #no 0.5*energy_tot
-end    
-
-"""
-    energy_update(i_atom, dist2_new, pot::AbstractDimerPotentialB)
-returns new energy when an atom is moved. In magnetuic field, without cutoff distance.
-"""
-function energy_update(i_atom, dist2_new, pot::AbstractDimerPotentialB)
-    return dimer_energy_atom(i_atom, dist2_new, pot)
-end
-
-"""
-    energy_update(i_atom, dist2_new, r_cut, pot::AbstractDimerPotentialB)
-returns new energy when an atom is moved. In magnetuic field, with cutoff distance r_cut.
-"""
-function energy_update(i_atom, dist2_new, r_cut, pot::AbstractDimerPotentialB)
-    return dimer_energy_atom(i_atom, dist2_new, r_cut, pot)
-end
-
-"""
-    energy_update(pos, i_atom, config, dist2_mat, tan_mat, pot::AbstractDimerPotentialB)
-Gives difference in energy, new distances, new tan when an atom is moved. In magnetic field, without cutoff distance.
-"""
-function energy_update(pos, i_atom, config, dist2_mat, tan_mat, pot::AbstractDimerPotentialB)
-    dist2_new = [distance2(pos,b,config.bc) for b in config.pos]
-    dist2_new[i_atom] = 0.
-    tan_new = [get_tan(pos,b,config.bc) for b in config.pos]
-    tan_new[i_atom] = 0.
-    d_en = dimer_energy_atom(i_atom, dist2_new, tan_new, pot) - dimer_energy_atom(i_atom, dist2_mat[:,i_atom], tan_mat[:,i_atom],pot)
-    return d_en, dist2_new, tan_new
-end
-
-"""
-    energy_update(pos, i_atom, config, dist2_mat, tan_mat, r_cut, pot::AbstractDimerPotentialB)
-Gives difference in energy, new distances, new tan when an atom is moved. In magnetic field, with cutoff distance r_cut.
-"""
-function energy_update(pos, i_atom, config, dist2_mat, tan_mat, r_cut, pot::AbstractDimerPotentialB)
-    dist2_new = [distance2(pos,b,config.bc) for b in config.pos]
-    dist2_new[i_atom] = 0.
-    tan_new = [get_tan(pos,b,config.bc) for b in config.pos]
-    tan_new[i_atom] = 0
-    d_en = dimer_energy_atom(i_atom, dist2_new, tan_new, r_cut, pot) - dimer_energy_atom(i_atom, dist2_mat[:,i_atom], tan_mat[:,i_atom],r_cut, pot)
-    return d_en, dist2_new, tan_new
-end
-"""
-    get_energy_dimer(pos,i_atom,mc_state,pot::AbstractDimerPotentialB)
-A get_energy function similar to the energy_update function. This simply returns the current energy rather than delta_en
-"""
-function get_energy_dimer(pos,i_atom,mc_state,pot::AbstractDimerPotentialB)
-    # dist2_new = [distance2(pos,b) for b in mc_state.config.pos]
-    # dist2_new[i_atom] = 0.
-    # delta_energy= dimer_energy_atom(i_atom, dist2_new, pot) - dimer_energy_atom(i_atom, mc_state.dist2_mat[:,i_atom], pot)
-
-    delta_energy,dist2_new,tan_new = energy_update(pos,i_atom,mc_state.config,mc_state.dist2_mat,mc_state.tan_mat,pot)
-    energy = mc_state.en_tot + delta_energy
-    return energy,[dist2_new,tan_new]
-end
-"""
-    get_energy_dimer(pos,i_atom,r_cut,mc_state,pot::AbstractDimerPotentialB)
-A get_energy function similar to the energy_update function. This simply returns the current energy rather than delta_en
-with a cutoff distance r_cut
-"""
-function get_energy_dimer(pos,i_atom,r_cut,mc_state,pot::AbstractDimerPotentialB)
-    # dist2_new = [distance2(pos,b) for b in mc_state.config.pos]
-    # dist2_new[i_atom] = 0.
-    # delta_energy= dimer_energy_atom(i_atom, dist2_new, pot) - dimer_energy_atom(i_atom, mc_state.dist2_mat[:,i_atom], pot)
-
-    delta_energy,dist2_new,tan_new = energy_update(pos,i_atom,mc_state.config,mc_state.dist2_mat,mc_state.tan_mat,r_cut,pot)
-    energy = mc_state.en_tot + delta_energy
-    return energy,[dist2_new,tan_new]
-end
-
-"""
-    get_energy(trial_positions,indices,mc_states,pot::AbstractDimerPotentialB)
-Top scope get energy function for dimer potentials returning the energy vector and new distance squared vector as this must be calculated in order to calculate the potential.
-
-"""
-function get_energy(trial_positions,indices,mc_states,pot::AbstractDimerPotentialB,ensemble::NVT)
-    energyvector, mats_new = invert(get_energy_dimer.(trial_positions,indices,mc_states,Ref(pot)))
-
-    # energyvector = mc_state.en_tot .+ delta_energyvector
-    return energyvector, mats_new
-end
-#this will be the format for this part of the get_energy function.
-
-"""
-    get_energy(trial_positions,indices,r_cut,mc_states,pot::AbstractDimerPotentialB,ensemble::NPT)
-Top scope get energy function for dimer potentials returning the energy vector and new distance squared vector as this must be calculated in order to calculate the potential.
-with the cutoff distance r_cut
-"""
-function get_energy(trial_positions,indices,mc_states,pot::AbstractDimerPotentialB,ensemble::NPT)
-    
-    n=length(mc_states)
-    r_cut_all=Array{Float64}(undef,n)
-    for i=1:n
-        r_cut_all[i]=mc_states[i].config.bc.box_length^2/4
-    end
-    energyvector, mats_new = invert(get_energy_dimer.(trial_positions,indices,r_cut_all,mc_states,Ref(pot)))
-    # energyvector = mc_state.en_tot .+ delta_energyvector
-    return energyvector,mats_new
-end
-
-"""
-    get_energy function when the whole configuration is scaled
-        find the new distance matrix first, and use dimer_energy_config to find the new total energy and energy matrix
-"""
-function get_energy(trial_configs_all,mc_states,pot::AbstractDimerPotentialB)
-    dist2_mat_new = get_distance2_mat.(trial_configs_all)
-    #en_atom_vec, en_tot_new = invert(dimer_energy_config.(dist2_mat_new, length(trial_configs_all[1].pos), Ref(pot)))
-    n=length(trial_configs_all)
-    r_cut_all=Array{Float64}(undef,n)
-    for i=1:n
-        r_cut_all[i]=trial_configs_all[i].bc.box_length^2/4
-    end
-    tan_mat=Array{Matrix}(undef,n)
-    for i=1:n
-        tan_mat[i]=mc_states[i].tan_mat
-    end
-    en_atom_vec, en_tot_new = invert(dimer_energy_config.(dist2_mat_new, tan_mat, length(trial_configs_all[1].pos), r_cut_all, Ref(pot)))
-    #println(invert(dimer_energy_config.(dist2_mat_new, length(trial_configs_all[1].pos), Ref(pot))))
-    #println()
-
-    return dist2_mat_new,en_atom_vec,en_tot_new
-end
-"""
-    end here
-"""
-
-
-""" 
-    DFTPotential 
-Implements type for a "density functional theory" potential (calcuate energies in DFT); subtype of AbstractPotential 
-field names: a: specifies the box length, lattice: specifies the 3x3 cube/box from a, El: specifies the atom type,
-pseudopotential and functional, atoms: a vector containing the atom type from El, functional: specifies the functional, 
-n_atoms:: specifies the number of atoms, kgrid: is the k-point sampling grid, Ecut: is energy cutoff. 
-""" 
-struct DFTPotential <:AbstractPotential
-    a::Float64                     
-    lattice::Mat3                  
-    El::ElementPsp                
-    atoms::Vector                 
-    functional::Vector{Symbol}    
-    n_atoms::Int                  
-    kgrid::Vector                 
-    Ecut::Int                      
-end  
-
-function DFTPotential(a, n_atoms) 
-    kgrid = [1, 1, 1] 
-    Ecut = 6
-    lattice = a * I(3) 
-    El = ElementPsp(:Ga, psp=load_psp("hgh/pbe/ga-q3")) 
-    atoms = Vector{ElementPsp}(undef,n_atoms)
-    for i in 1:n_atoms 
-        atoms[i] = El 
-    end  
-    functional = [:gga_x_pbe, :gga_c_pbe] 
-    return DFTPotential(a, lattice, El, atoms, functional, n_atoms, kgrid, Ecut)
-end  
-""" 
-    getenergy_DFT(pos1, pot) 
-Calculates total energy of a given configuration for an arbitrary number of gallium atoms; 
-note that this function depends only on the positions of the atoms within the configuration, 
-so no bc's are to be included. 
-"""
-function getenergy_DFT(pos1, pot::DFTPotential) 
-    pos1 = pos1 / pot.a 
-    model = model_DFT(pot.lattice, pot.atoms, pos1, pot.functional)
-    basis = PlaneWaveBasis(model; pot.Ecut, pot.kgrid) 
-    scfres = self_consistent_field(basis; tol = 1e-7, callback=info->nothing) 
-    return scfres.energies.total 
-end  
-
-function energy_update(pos, i_atom, config::Config, dist2_mat, en_old, pot::DFTPotential) #pos is SVector, i_atom is integer 
-    dist2_new = [distance2(pos,b) for b in config.pos]
-    dist2_new[i_atom] = 0.  
-    config.pos[i_atom] = copy(pos)
-    pos_new = copy(config.pos) 
-    delta_en = getenergy_DFT(pos_new, pot) - en_old
-    return delta_en, dist2_new
-end   
-
-#--------------------------------------------#
-#--------------RuNNer methods----------------#
 """
     RuNNerPotential <: AbstractMachineLearningPotential
 Contains the important structs required for a neural network potential defined in the MachineLearningPotential package:
@@ -899,87 +434,186 @@ function RuNNerPotential(nnp,symmetryfunction)
     r_cut = symmetryfunction[1].r_cut
     return RuNNerPotential(nnp,symmetryfunction,r_cut)
 end
+
+
+mutable struct NNPVariables <: PotentialVariables
+    en_atom_vec::Vector
+    new_en::Float64
+    new_en_atom::Vector
+    g_matrix::Array
+    f_matrix::Array
+    new_g_matrix::Array
+    new_f_vec::Vector
+end
 """
     get_new_state_vars!(trial_pos,atomindex,nnp_state,pot)
-calculates the changes to the g_matrix, dist2_matrix etc based on the new `trial_pos` of atom `atomindex` and adjusts these values inside the `nnp_state` struct. Note the use of the threaded function for the calculation of the symmetry functions
+Function for finding the new state variables for calculating an NNP. Redefines new_f and new_g matrices based on the `trial_pos` of atom at `atomindex` and adjusts the parameters in the `nnp_state` according to the variables in `pot`.
 """
 function get_new_state_vars!(trial_pos,atomindex,nnp_state,pot)
 
-    nnp_state.new_dist2_vec = [ distance2(trial_pos,b,nnp_state.config.bc) for b in nnp_state.config.pos]
-    nnp_state.new_dist2_vec[atomindex] = 0.
+    nnp_state.potential_variables.new_dist2_vec = [ distance2(trial_pos,b,nnp_state.config.bc) for b in nnp_state.config.pos]
+    nnp_state.potential_variables.new_dist2_vec[atomindex] = 0.
     
-    nnp_state.new_f_vec = cutoff_function.(sqrt.(nnp_state.new_dist2_vec),Ref(pot.r_cut))
+    nnp_state.potential_variables.new_f_vec = cutoff_function.(sqrt.(nnp_state.potential_variables.new_dist2_vec),Ref(pot.r_cut))
 
 
-    nnp_state.new_g_matrix = copy(nnp_state.g_matrix)
+    nnp_state.potential_variables.new_g_matrix = copy(nnp_state.potential_variables.g_matrix)
 
-    nnp_state.new_g_matrix = total_thr_symm!(nnp_state.new_g_matrix,nnp_state.config.pos,trial_pos,nnp_state.dist2_mat,nnp_state.new_dist2_vec,nnp_state.f_matrix,nnp_state.new_f_vec,atomindex,pot.symmetryfunctions)
+    nnp_state.potential_variables.new_g_matrix = total_thr_symm!(nnp_state.potential_variables.new_g_matrix,nnp_state.config.pos,trial_pos,nnp_state.dist2_mat,nnp_state.potential_variables.new_dist2_vec,nnp_state.potential_variables.f_matrix,nnp_state.potential_variables.new_f_vec,atomindex,pot.symmetryfunctions)
 
     return nnp_state
 end
-
 """
-    calc_new_energy!(nnp_state,pot)
-Function to curry the PTMC struct `nnp_state` and `pot` into the format required by the MachineLearningPotential package, output is the updated `nnp_state`
+    calc_new_runner_energy!(potential_variables,pot)
+function designed to calculate the new per-atom energy according to the RuNNer forward pass with parameters defined in `pot`. utilises the `new_g_matrix` to redefine the `new_en` and `new_en_atom` variables within the `potential_variables` struct.
 """
-function calc_new_energy!(nnp_state,pot)
-    nnp_state.new_en_atom = forward_pass(nnp_state.new_g_matrix,length(nnp_state.en_atom_vec),pot.nnp)
-    return nnp_state
+function calc_new_runner_energy!(mc_state,pot)
+    mc_state.new_en_atom = forward_pass(mc_state.potential_variables.new_g_matrix,length(mc_state.en_atom_vec),pot.nnp)
+    mc_state.new_en = sum(mc_state.potential_variables.new_en_atom)
+    return mc_state
 end
-"""
-    get_runner_energy!(trial_pos,atomindex,nnp_state,pot)
-base-level single-state calculation of the `energy` and adjusted `nnp_state` given that we have displaced atom `atomindex` to position `trial_pos`.
-"""
-function get_runner_energy!(trial_pos,atomindex,nnp_state,pot)
-    nnp_state = get_new_state_vars!(trial_pos,atomindex,nnp_state,pot)
 
-    nnp_state = calc_new_energy!(nnp_state,pot)
+
+#----------------------------------------------------------#
+#----------------------Top Level Call----------------------#
+#----------------------------------------------------------#
+"""
+    energy_update(trial_pos,index,mc_state,pot::AbstractDimerPotential)
+    energy_update!(trial_pos,index,mc_state,pot::AbstractDimerPotentialB)
+    energy_update!(trial_pos,atomindex,mc_state,pot::EmbeddedAtomPotential)
+    energy_update!(trial_pos,index,state,pot::RuNNerPotential)
+Energy update function for use within a cycle. at the top level this is called with the new position `trial_pos` which is the `index`-th atom in the `config` contained in `mc_state`. Using `pot` the potential. 
     
-    return sum(nnp_state.new_en_atom) , nnp_state
-end
-"""
-    get_energy!(trial_positions,indices,nnp_states,pot::RuNNerPotential)
-Total, vectorised function for the calculation of the energy change in each state within `nnp_states` based on a vector of moved `indices` to `trial_positions` based on the potential `pot`.
-"""
-function get_energy(trial_positions,indices,nnp_states,pot::RuNNerPotential)
-    energyvector = Vector{Float64}(undef,length(indices))
+    This function is designed as a curry function. The generic get_energy function operates on a __vector__ of states, this function takes each state and the set potential and calls the potential specific energy_update function.
 
-    energyvector,nnp_states = invert(get_runner_energy!.(trial_positions,indices,nnp_states,Ref(pot)))
+        Methods defined for :
+            - AbstractDimerPotential
+            - AbstractDimerPotentialB
+            - EmbeddedAtomPotential
+            - RuNNerPotential
+"""
+function energy_update!(trial_pos,index,mc_state,pot::AbstractDimerPotential)
+
+    mc_state.new_dist2_vec = [distance2(trial_pos,b,mc_state.config.bc) for b in mc_state.config.pos]
+    mc_state.new_dist2_vec[index] = 0.
+
+    mc_state.new_en = dimer_energy_update!(index,mc_state.dist2_mat,mc_state.new_dist2_vec,mc_state.en_tot,pot)
+
+    return mc_state
+end
+function energy_update!(trial_pos,index,mc_state,pot::AbstractDimerPotentialB)
+
+    mc_state.new_dist2_vec = [distance2(trial_pos,b,mc_state.config.bc) for b in mc_state.config.pos]
+    mc_state.new_dist2_vec[index] = 0.
+
+    mc_state.potential_variables.new_tan_vec = [get_tan(trial_pos,b,mc_state.config.bc) for b in mc_state.config.pos]
+    mc_state.potential_variables.new_tan_vec[index] = 0
+
+    mc_state.new_en = dimer_energy_update!(index,mc_state.dist2_mat,mc_state.potential_variables.tan_mat,mc_state.new_dist2_vec,mc_state.potential_variables.new_tan_vec,mc_state.en_tot,pot)
+
+    return mc_state
+end
+function energy_update!(trial_pos,atomindex,mc_state,pot::EmbeddedAtomPotential)
+    mc_state.new_dist2_vec = [distance2(trial_pos,b) for b in mc_state.config.pos]
+
+    mc_state.new_dist2_vec[atomindex] = 0.
+
+    mc_state.potential_variables.new_component_vector = deepcopy(mc_state.potential_variables.component_vector)
     
-    return energyvector, nnp_states
-end
+    mc_state.potential_variables.new_component_vector = calc_components(mc_state.potential_variables.new_component_vector,atomindex,mc_state.dist2_mat[atomindex,:],mc_state.new_dist2_vec,pot.n,pot.m)
 
-#---------------------------------------------#
+    mc_state.new_en = calc_energies_from_components(mc_state.potential_variables.new_component_vector,pot.ean,pot.eCam)
+
+    return  mc_state
+end
+function energy_update!(trial_pos,index,mc_state,pot::RuNNerPotential)
+
+    mc_state = get_new_state_vars(trial_pos,index,mc_state,pot)
+
+    mc_state = calc_new_runner_energy!(mc_state,pot)
+    return mc_state
+end
 
 """
-    EnHist(n_bin, en_min::T, en_max::T)
-    EnHist(n_bin; en_min=-0.006, en_max=-0.001)
-Collects data for energy histograms per temperature
-Field names:    
-- `n_bins`: number of energy bins
-- `en_min`,`en_max`: minimum and maximum energy between which data is collected
-- `delta_en_bin`: energy spacing of bins
-- `en_hist`: stores number of sampled configurations per energy bins
+    get_energy(trial_positions,indices,mc_states,pot)
+    get_energy(trial_positions,indices,mc_states,pot::RuNNerPotential)
+curry function used as the top call within each mc_step. Passes a vector of `mc_states` with updated `trial_positions` for atom at `index` where we use `pot` to calculate the new energy in the non-vectorised energy_update function. 
+
 """
-struct EnHist{T}
-    n_bin::Int
-    en_min::Ref{T}
-    en_max::Ref{T}
-    delta_en_bin::Ref{T}
-    en_hist::Vector{Int}
+
+function get_energy(trial_positions,indices,mc_states,pot)
+    return energy_update!.(trial_positions,indices,mc_states,Ref(pot))
 end
 
-function EnHist(n_bin, en_min::T, en_max::T) where T
-    delta_en_bin = (en_max-en_min)/n_bin
-    en_hist = zeros(Int, n_bin)
-    return EnHist{T}(n_bin,en_min,en_max,delta_en_bin,en_hist)
+
+#------------------------------------------------------------#
+#----------------Initialising State Functions----------------#
+#------------------------------------------------------------#
+"""
+    initialise_energy(config,dist2_mat,potential_variables,pot::AbstractDimerPotential)
+    initialise_energy(config,dist2_mat,potential_variables,r_cut,pot::AbstractDimerPotential)
+    initialise_energy(config,dist2_mat,potential_variables,r_cut,pot::RuNNerPotential)
+
+Initialise energy is used during the MCState call to set the starting energy of a `config` according to the potential as `pot` and the configurational variables `potential_variables`. Written with general input means the top-level is type-invariant. 
+Methods included for:
+    - Dimer Potential 
+    - Machine Learning Potentials 
+"""
+function initialise_energy(config,dist2_mat,potential_variables,pot::AbstractDimerPotential)
+    potential_variables.en_atom_vec,en_tot = dimer_energy_config(dist2_mat,length(config),pot)
+
+    return en_tot,potential_variables
+end
+function initialise_energy(config,dist2_mat,potential_variables,pot::AbstractDimerPotentialB)
+    potential_variables.en_atom_vec,en_tot = dimer_energy_config(dist2_mat,potential_variables.tan_mat,length(config),pot)
+    return en_tot,potential_variables 
+end
+function initialise_energy(config,dist2_mat,potential_variables,pot::EmbeddedAtomPotential)
+    en_tot = calc_energies_from_components(potential_variables.component_vector,pot.ean,pot.eCam)
+
+    return en_tot,potential_variables
+end
+function initialise_energy(config,dist2_mat,potential_variables,pot::RuNNerPotential)
+    potential_variables.en_atom_vec = forward_pass(potential_variables.g_matrix,length(config),pot.nnp)
+    en_tot = sum(potential_variables.en_atom_vec)
+    return en_tot,potential_variables
 end
 
-function EnHist(n_bin; en_min=-0.006, en_max=-0.001)
-    T = eltype(en_min)
-    delta_en_bin = (en_max-en_min)/n_bin
-    en_hist = zeros(Int, n_bin)
-    return EnHist{T}(n_bin,en_min,en_max,delta_en_bin,en_hist)
+"""
+    set_variables(config,dist_2_mat,pot::AbstractDimerPotential)
+    set_variables(config,dist2_matrix,pot::EmbeddedAtomPotential)
+    set_variables(config,dist_2_mat,pot::AbstractDimerPotential)
+
+initialises the PotentialVariable struct for the various potentials. 
+    -   Method one functions for abstract dimer potentials such as the ELJ
+    -   Embeded Atom Potential involving the dimer and density components
+    -   The RuNNer Potential involving the cutoff and symmetry matrix.
+    
+"""
+function set_variables(config,dist_2_mat,pot::AbstractDimerPotential)
+    return DimerPotentialVariables()
+end
+function set_variables(config::Config,dist2_matrix::Matrix,pot::AbstractDimerPotentialB)
+    n_atoms = length(config)
+    tan_matrix = get_tantheta_mat(config,config.pos)
+
+    return ELJPotentialBVariables(tan_matrix,zeros(n_atoms))
+end
+function set_variables(config,dist2_matrix,pot::EmbeddedAtomPotential)
+    n_atoms = length(config)
+    componentvec = zeros(n_atoms,2)
+    for row_index in 1:n_atoms
+        componentvec[row_index,:] = calc_components(componentvec[row_index,:],dist2_matrix[row_index,:],pot.n,pot.m)
+    end
+    return EmbeddedAtomVariables(componentvec,zeros(n_atoms,2))
+end
+function set_variables(config,dist2_mat,pot::RuNNerPotential)
+    
+    n_atoms = length(config)
+    f_matrix = cutoff_function.(sqrt.(dist2_mat),Ref(pot.r_cut))
+    g_matrix = total_symm_calc(config.pos,dist2_mat,f_matrix,pot.symmetryfunctions)
+    
+    return NNPVariables(0. ,zeros(n_atoms),g_matrix,f_matrix,zeros(length(pot.symmetryfunctions)), zeros(n_atoms))
 end
 
 end
